semantic-search/pyproject.toml
[project]
name = "semantic-search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-community>=0.3.29",
    "langchain-core>=0.3.75",
    "langchain-openai>=0.3.32",
    "load-dotenv>=0.1.0",
    "pypdf>=6.0.0",
]
